feat: implement rate limiting. refactor: defined securityWrappers for middleware.js
