+++ package.json
@@ -17,6 +17,7 @@
     },
     "dependencies": {
         "@google-cloud/storage": "^7.19.0",
+        "@upstash/ratelimit": "^2.0.8",
         "bootstrap": "^5.3.8",
         "browser-image-compression": "^2.0.2",
         "date-fns": "^3.6.0",
