+++ src/app/api/microsoft/msGraphFunctions.js
@@ -7,7 +7,7 @@
  * Only import from other API routes (server-side).
  */
 
-import { sanitiseHtml } from '@/lib/security/securityHelpers';
+import { sanitiseHtml } from '@/lib/security/securityWrappers';
 
 const GRAPH_BASE_URL = 'https://graph.microsoft.com/v1.0';
 const TIMEZONE = 'Australia/Sydney';
