--- src/lib/security/securityHelpers.js
@@ -1,7 +1,0 @@
-import sanitizeHtmlLibrary from 'sanitize-html';
-
-export const sanitiseHtml = (html) =>
-    sanitizeHtmlLibrary(html, {
-        allowedTags: sanitizeHtmlLibrary.defaults.allowedTags,
-        allowedAttributes: sanitizeHtmlLibrary.defaults.allowedAttributes,
-});
